+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps/Logics/DefaultLogic.cs
@@ -3,11 +3,12 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 
-namespace Xamarin.Forms.GoogleMaps.Logics.iOS
+namespace Xamarin.Forms.GoogleMaps.Logics
 {
-    internal abstract class DefaultLogic<TOuter, TNative> : BaseLogic
+    internal abstract class DefaultLogic<TOuter, TNative, TNativeMap> : BaseLogic<TNativeMap>
         where TOuter : BindableObject
         where TNative : class
+        where TNativeMap : class
     {
         readonly IList<TOuter> _outerItems = new List<TOuter>(); // Only for ResetItems.
